cuAndSort
