+++ src/dsstats.mmr/MmrService.Calc.cs
@@ -1,8 +1,8 @@
-
 using dsstats.mmr.Extensions;
 using dsstats.mmr.ProcessData;
 using pax.dsstats.shared;
 using pax.dsstats.shared.Raven;
+
 using TeamData = dsstats.mmr.ProcessData.TeamData;
 
 namespace dsstats.mmr;
@@ -147,93 +147,6 @@ public static partial class MmrService
         return changes;
     }
 
-    //# Commander-Mmr
-    private static double GetCommandersComboMmr(ReplayData replayData, TeamData teamData, Dictionary<CmdrMmmrKey, CmdrMmmrValue> cmdrMmrDic)
-    {
-        if (replayData.IsInvalid)
-        {
-            return 1.0;
-        }
-
-        double commandersComboMMRSum = 0;
-
-        for (int playerIndex = 0; playerIndex < teamData.Players.Length; playerIndex++)
-        {
-            var playerCmdr = teamData.Players[playerIndex].Race;
-
-            double synergySum = 0;
-            double antiSynergySum = 0;
-
-            for (int synergyPlayerIndex = 0; synergyPlayerIndex < teamData.Players.Length; synergyPlayerIndex++)
-            {
-                if (playerIndex == synergyPlayerIndex)
-                {
-                    continue;
-                }
-
-                var synergyPlayerCmdr = teamData.Players[synergyPlayerIndex].Race;
-
-                var synergy = cmdrMmrDic[new CmdrMmmrKey() { Race = playerCmdr, OppRace = synergyPlayerCmdr }];
-
-                synergySum += ((1 / 2.0) * synergy.SynergyMmr);
-            }
-
-            for (int antiSynergyPlayerIndex = 0; antiSynergyPlayerIndex < teamData.Players.Length; antiSynergyPlayerIndex++)
-            {
-                var antiSynergyPlayerCmdr = teamData.Players[antiSynergyPlayerIndex].OppRace;
-
-                var antiSynergy = cmdrMmrDic[new CmdrMmmrKey() { Race = playerCmdr, OppRace = antiSynergyPlayerCmdr }];
-
-                if (playerIndex == antiSynergyPlayerIndex)
-                {
-                    antiSynergySum += (MmrOptions.ownMatchupPercentage * antiSynergy.AntiSynergyMmr);
-                }
-                else
-                {
-                    antiSynergySum += (MmrOptions.matesMatchupsPercentage * antiSynergy.AntiSynergyMmr);
-                }
-            }
-
-            commandersComboMMRSum +=
-                (MmrOptions.antiSynergyPercentage * antiSynergySum)
-                + (MmrOptions.synergyPercentage * synergySum);
-        }
-
-        return commandersComboMMRSum / 3;
-    }
-
-    private static void SetCommandersComboMmr(ReplayData replayData, TeamData teamData, Dictionary<CmdrMmmrKey, CmdrMmmrValue> cmdrMmrDic)
-    {
-        if (replayData.IsInvalid) {
-            return;
-        }
-
-        for (int playerIndex = 0; playerIndex < teamData.Players.Length; playerIndex++) {
-            var playerCmdr = teamData.Players[playerIndex].Race;
-
-            for (int synergyPlayerIndex = 0; synergyPlayerIndex < teamData.Players.Length; synergyPlayerIndex++) {
-                if (playerIndex == synergyPlayerIndex) {
-                    continue;
-                }
-
-                var synergyPlayerCmdr = teamData.Players[synergyPlayerIndex].Race;
-
-                var synergy = cmdrMmrDic[new CmdrMmmrKey() { Race = playerCmdr, OppRace = synergyPlayerCmdr }];
-
-                synergy.SynergyMmr += teamData.Players[playerIndex].Deltas.CommanderMmr / 2;
-            }
-
-            for (int antiSynergyPlayerIndex = 0; antiSynergyPlayerIndex < teamData.Players.Length; antiSynergyPlayerIndex++) {
-                var antiSynergyPlayerCmdr = teamData.Players[antiSynergyPlayerIndex].OppRace;
-
-                var antiSynergy = cmdrMmrDic[new CmdrMmmrKey() { Race = playerCmdr, OppRace = antiSynergyPlayerCmdr }];
-
-                antiSynergy.AntiSynergyMmr += teamData.Players[playerIndex].Deltas.CommanderMmr;
-            }
-        }
-    }
-
-
     //# Set ProcessData
     private static void SetReplayData(Dictionary<int, CalcRating> mmrIdRatings,
                                       ReplayData replayData,
@@ -297,40 +210,5 @@ public static partial class MmrService
         playerData.Mmr = plRating.Mmr;
         playerData.Consistency = plRating.Consistency;
         playerData.Confidence = plRating.Confidence;
-    }
-
-
-    //# Formulas
-    private static double GetCorrectedConfidenceFactor(double playerConfidence, double replayConfidence)
-    {
-        double totalConfidenceFactor = (0.5 * (1 - GetConfidenceFactor(playerConfidence))) + (0.5 * GetConfidenceFactor(replayConfidence));
-        return 1 + MmrOptions.confidenceImpact * (totalConfidenceFactor - 1);
-    }
-
-    private static double GetConfidenceFactor(double confidence)
-    {
-        double variance = ((MmrOptions.distributionMult * 0.4) + (1 - confidence));
-
-        return MmrOptions.distributionMult * (1 / (Math.Sqrt(2 * Math.PI) * Math.Abs(variance)));
-    }
-
-    private static double CalculateMmrDelta(double elo, double playerImpact, double eloK = 32)
-    {
-        return (double)(eloK * (1 - elo) * playerImpact);
-    }
-
-    private static double GetCorrectedRevConsistency(double raw_revConsistency)
-    {
-        return 1 + MmrOptions.consistencyImpact * (raw_revConsistency - 1);
-    }
-
-    private static double PlayerToTeamMates(double teamMmrMean, double playerMmr, int teamSize)
-    {
-        return teamSize * (playerMmr / (teamMmrMean * teamSize));
-    }
-
-    public static double EloExpectationToWin(double ratingOne, double ratingTwo, double clip = 400)
-    {
-        return 1.0 / (1.0 + Math.Pow(10.0, (2.0 / clip) * (ratingTwo - ratingOne)));
     }
 }
